pulls campaign and displays title
